Implement a method to check the status code of an `api.ErrorResponse`
This is a very common use case, but the error unwrapping pattern is
pretty tedious to write each time, and the resulting line of code is usually
very long.

The pattern itself is also easy to get wrong as you need to pass a pointer
to a pointer to the second argument of `errors.As`, but it accepts
a value of `any` type.

The combination of a very long line and code that needs to be checked
carefully is precarious, so I think abstracting this pattern is worth
it.